Rename p2p_types to just types
