+++ elements/lisk-p2p/src/peer/outbound.ts
@@ -40,7 +40,7 @@ import {
 	P2PPeerInfo,
 	P2PRequestPacket,
 	P2PResponsePacket,
-} from '../p2p_types';
+} from '../types';
 
 import * as querystring from 'querystring';
 import * as socketClusterClient from 'socketcluster-client';
