+++ elements/lisk-p2p/src/peer_directory/peer_book.ts
@@ -13,7 +13,7 @@
  *
  */
 
-import { P2PDiscoveredPeerInfo, P2PPeerInfo } from '../p2p_types';
+import { P2PDiscoveredPeerInfo, P2PPeerInfo } from '../types';
 import { NewPeerConfig, NewPeers } from './new_peers';
 import { TriedPeerConfig, TriedPeers } from './tried_peers';
