+++ elements/lisk-p2p/src/peer_directory/tried_peers.ts
@@ -17,7 +17,7 @@ import {
 	DEFAULT_TRIED_PEER_BUCKET_SIZE,
 	DEFAULT_TRIED_PEER_LIST_SIZE,
 } from '..';
-import { P2PDiscoveredPeerInfo, P2PPeerInfo } from '../p2p_types';
+import { P2PDiscoveredPeerInfo, P2PPeerInfo } from '../types';
 import { constructPeerIdFromPeerInfo, getBucket, PEER_TYPE } from '../utils';
 
 export interface TriedPeerConfig {
